Extract Direct Link generator underneath Builder namespace
