+++ lib/peanut_labs.rb
@@ -3,6 +3,6 @@ require 'peanut_labs/builder/user_id'
 require 'peanut_labs/builder/profile'
 require 'peanut_labs/errors'
 require 'peanut_labs/builder/iframe_url'
-require 'peanut_labs/direct_link'
+require 'peanut_labs/builder/direct_link'
 
 module PeanutLabs; end
